+++ NEWS.markdown
@@ -1,5 +1,5 @@
 <!--
-SPDX-FileCopyrightText: Copyright (c) 2023 Paul A. Patience <paul@apatience.com>
+SPDX-FileCopyrightText: Copyright (c) 2023-2024 Paul A. Patience <paul@apatience.com>
 SPDX-License-Identifier: MIT
 -->
 
@@ -12,6 +12,12 @@ adheres to [Semantic Versioning][].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.1.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
+
+## Unreleased
+
+### Fixed
+
+- Removed iteration clause from LOOP body code.
 
 ## 0.1.0 — 2023-12-14